Commit 6
Fixed almost everything the reviewer asked to fix. 
Got stuck with with the textFields. After refactoring the code within viewDidLoad both textFields aligned to the left. 
I know that texfields aligned to the left by default in Xcode but I do have code aligning them in the centre so what the hell?!